refactor : 코드 병합
