+++ src/main/java/com/example/randomdriveproject/controller/HomeController.java
@@ -1,6 +1,6 @@
 package com.example.randomdriveproject.controller;
 
-import com.example.randomdriveproject.service.RandomUserService;
+import com.example.randomdriveproject.user.service.RandomUserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
