+++ src/main/java/com/example/randomdriveproject/controller/RouteController.java
@@ -1,7 +1,7 @@
 package com.example.randomdriveproject.controller;
 
-import com.example.randomdriveproject.dto.kakao_navigation.KakaoRouteAllResponseDto;
-import com.example.randomdriveproject.service.RandomkakaoRouteSearchService;
+import com.example.randomdriveproject.navigation.basic.service.KakaoRouteSearchService;
+import com.example.randomdriveproject.request.dto.KakaoRouteAllResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 @RequiredArgsConstructor
 public class RouteController {
 
-    private final RandomkakaoRouteSearchService kakaoRouteSearchService;
+    private final KakaoRouteSearchService kakaoRouteSearchService;
 
     @Value("${kakao.javascript.api.key}")
     private String kakaoJavascriptApiKey;
