--- src/main/java/com/example/randomdriveproject/randomNavi/api/service/RandomKakaoAddressSearchService.java
@@ -1,43 +1,0 @@
-package com.example.randomdriveproject.randomNavi.api.service;
-
-import com.example.randomdriveproject.randomNavi.api.dto.RandomKakaoApiResponseDto;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.stereotype.Service;
-import org.springframework.util.ObjectUtils;
-import org.springframework.web.client.RestTemplate;
-
-import java.net.URI;
-
-@Slf4j(topic = "KakaoAddressSearchService")
-@Service
-@RequiredArgsConstructor
-public class RandomKakaoAddressSearchService {
-
-    private final RestTemplate restTemplate;
-    private final RandomKakaoUriBuilderService kakaoUriBuilderService;
-
-    @Value("${kakao.rest.api.key}")
-    private String kakaoRestApiKey;
-
-    // 해당 메서드를 호출 -> 입력한 주소값 -> 요청할 수 있는 uri + headers에 key값
-    // 요청한 것에 대한 응답 값 -> kakaoApiResponseDto 객체로 반환
-    public RandomKakaoApiResponseDto requestAddressSearch(String address){ // 주소 -> 위도, 경도로 변환
-
-        if(ObjectUtils.isEmpty(address)) return null;
-
-        URI uri = kakaoUriBuilderService.buildUriByAddressSearch(address);
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.set(HttpHeaders.AUTHORIZATION, "KakaoAK " + kakaoRestApiKey);
-        HttpEntity httpEntity = new HttpEntity(headers);
-
-        // kakao api 호출
-        return restTemplate.exchange(uri, HttpMethod.GET, httpEntity, RandomKakaoApiResponseDto.class).getBody();
-
-    }
-}
