+++ src/main/java/com/example/randomdriveproject/request/dto/KakaoApiResponseDto.java
@@ -1,0 +1,20 @@
+package com.example.randomdriveproject.request.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class KakaoApiResponseDto {
+
+    @JsonProperty("meta")
+    private MetaDto metaDto;
+
+    @JsonProperty("documents")
+    private List<DocumentDto> documentDtoList;
+}
