+++ src/main/java/com/example/randomdriveproject/request/dto/MetaDto.java
@@ -1,0 +1,14 @@
+package com.example.randomdriveproject.request.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class MetaDto {
+
+    @JsonProperty("total_count")
+    private Integer totalCount;
+
+}
