+++ src/main/java/com/example/randomdriveproject/user/dto/kakao/KakaoProfile.java
@@ -1,0 +1,15 @@
+package com.example.randomdriveproject.user.dto.kakao;
+
+import lombok.Getter;
+
+import java.util.Map;
+
+@Getter
+public class KakaoProfile {
+    private String nickname;
+
+    public KakaoProfile(Map<String, Object> data)
+    {
+        nickname = (String) data.get("nickname");
+    }
+}
